RUM-9309: Add `ActionTrackingStrategy` Interface to decouple find view logic
