+++ features/dd-sdk-android-rum/api/dd-sdk-android-rum.api
@@ -6777,6 +6777,11 @@ public class com/datadog/android/rum/tracking/AcceptAllSupportFragments : com/da
 	public fun hashCode ()I
 }
 
+public abstract interface class com/datadog/android/rum/tracking/ActionTrackingStrategy {
+	public abstract fun findTargetForScroll (Landroid/view/View;FF)Lcom/datadog/android/rum/tracking/ViewTarget;
+	public abstract fun findTargetForTap (Landroid/view/View;FF)Lcom/datadog/android/rum/tracking/ViewTarget;
+}
+
 public abstract class com/datadog/android/rum/tracking/ActivityLifecycleTrackingStrategy : android/app/Application$ActivityLifecycleCallbacks, com/datadog/android/rum/tracking/TrackingStrategy {
 	protected field sdkCore Lcom/datadog/android/api/feature/FeatureSdkCore;
 	public fun <init> ()V
@@ -6858,6 +6863,21 @@ public abstract interface class com/datadog/android/rum/tracking/TrackingStrateg
 
 public abstract interface class com/datadog/android/rum/tracking/ViewAttributesProvider {
 	public abstract fun extractAttributes (Landroid/view/View;Ljava/util/Map;)V
+}
+
+public final class com/datadog/android/rum/tracking/ViewTarget {
+	public fun <init> ()V
+	public fun <init> (Landroid/view/View;Ljava/lang/String;)V
+	public synthetic fun <init> (Landroid/view/View;Ljava/lang/String;ILkotlin/jvm/internal/DefaultConstructorMarker;)V
+	public final fun component1 ()Landroid/view/View;
+	public final fun component2 ()Ljava/lang/String;
+	public final fun copy (Landroid/view/View;Ljava/lang/String;)Lcom/datadog/android/rum/tracking/ViewTarget;
+	public static synthetic fun copy$default (Lcom/datadog/android/rum/tracking/ViewTarget;Landroid/view/View;Ljava/lang/String;ILjava/lang/Object;)Lcom/datadog/android/rum/tracking/ViewTarget;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getTag ()Ljava/lang/String;
+	public final fun getView ()Landroid/view/View;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
 }
 
 public abstract interface class com/datadog/android/rum/tracking/ViewTrackingStrategy : com/datadog/android/rum/tracking/TrackingStrategy {
