+++ features/dd-sdk-android-rum/src/main/kotlin/com/datadog/android/rum/tracking/ActionTrackingStrategy.kt
@@ -1,0 +1,40 @@
+/*
+ * Unless explicitly stated otherwise all files in this repository are licensed under the Apache License Version 2.0.
+ * This product includes software developed at Datadog (https://www.datadoghq.com/).
+ * Copyright 2016-Present Datadog, Inc.
+ */
+
+package com.datadog.android.rum.tracking
+
+import android.view.View
+import com.datadog.tools.annotation.NoOpImplementation
+
+/**
+ * Strategy interface for tracking user interaction targets within a UI,
+ * such as taps and scrolls.
+ */
+@NoOpImplementation
+interface ActionTrackingStrategy {
+
+    /**
+     * Finds the target element at the given screen coordinates in response to a tap action.
+     *
+     * @param decorView The root [View] of the current window hierarchy.
+     * @param x The X coordinate of the tap, in screen coordinates.
+     * @param y The Y coordinate of the tap, in screen coordinates.
+     * @return A [ViewTarget] representing the UI element found at the given location,
+     *         or `null` if no target was found.
+     */
+    fun findTargetForTap(decorView: View, x: Float, y: Float): ViewTarget?
+
+    /**
+     * Finds the target element at the given screen coordinates in response to a scroll action.
+     *
+     * @param decorView The root [View] of the current window hierarchy.
+     * @param x The X coordinate of the scroll event, in screen coordinates.
+     * @param y The Y coordinate of the scroll event, in screen coordinates.
+     * @return A [ViewTarget] representing the UI element found at the given location,
+     *         or `null` if no target was found.
+     */
+    fun findTargetForScroll(decorView: View, x: Float, y: Float): ViewTarget?
+}
